+++ lib/epoch.ex
@@ -1,4 +1,8 @@
 defmodule Epoch do
+  @type epoch :: integer
+  @type epoch_in_ms :: String.t()
+  @type datetime :: DateTime.t()
+
   @moduledoc """
   Conversion between Elixir native `DateTime.t/0` and epoch formats.
   """
@@ -6,7 +10,7 @@ defmodule Epoch do
   @doc """
   Returns the current time in epoch format.
   """
-  @spec now :: integer
+  @spec now :: epoch
   def now do
     DateTime.utc_now() |> DateTime.to_unix()
   end
@@ -14,7 +18,7 @@ defmodule Epoch do
   @doc """
   Returns the current time in epoch format with millisecond resolution.
   """
-  @spec now(:microsecond) :: String.t()
+  @spec now(:microsecond) :: epoch_in_ms
   def now(:microsecond) do
     cast(DateTime.utc_now())
   end
@@ -31,10 +35,27 @@ defmodule Epoch do
       "1608431189.000000"
 
   """
-  @spec cast(DateTime.t()) :: String.t()
+  @spec cast(datetime) :: epoch_in_ms
   def cast(%DateTime{} = time) do
     epoch_in_ms = DateTime.to_unix(time, :microsecond)
     :erlang.float_to_binary(epoch_in_ms / 1_000_000.00, decimals: 6)
+  end
+
+  @doc """
+  Returns seconds since epoch for the given `DateTime.t/0`.
+
+  ## Examples
+
+      iex> seconds_since(~U[2020-12-20 02:26:29.000800Z])
+      1608431189
+
+      iex> seconds_since(~U[2020-12-20 02:26:29Z])
+      1608431189
+
+  """
+  @spec seconds_since(datetime) :: epoch
+  def seconds_since(%DateTime{} = time) do
+    DateTime.to_unix(time, :second)
   end
 
   @doc """
@@ -52,7 +73,7 @@ defmodule Epoch do
       ~U[2020-12-20 02:26:29Z]
 
   """
-  @spec to_datetime(String.t() | integer()) :: DateTime.t()
+  @spec to_datetime(epoch_in_ms | epoch) :: datetime
   def to_datetime(epoch) when is_binary(epoch) do
     if String.contains?(epoch, ".") do
       epoch